+++ tap_chargify/__init__.py
@@ -89,4 +89,3 @@ def main():
         sync(client, parsed_args.catalog, state)
 
 
-